requirements.txt
# Social Media Automation Engine — Dependencies

# Claude AI
anthropic>=0.25.0

# X (Twitter) API
tweepy>=4.14.0

# Reddit API
praw>=7.7.1

# Telegram Bot
python-telegram-bot>=20.7

# Task scheduler (Fase 2)
APScheduler>=3.10.4

# Environment variables
python-dotenv>=1.0.0

# HTTP utilities
httpx>=0.27.0
tenacity>=8.2.3
